requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
biopython==1.79
blis==0.7.7
blosc @ file:///C:/ci_310/python-blosc_1642117183052/work
cadences==0.3.4
catalogue==2.0.7
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
clldutils==3.12.0
colorama==0.4.4
colorlog==6.6.0
csvw==2.0.0
cycler==0.11.0
cymem==2.0.6
Cython==0.29.28
datasets==2.2.2
decorator==4.4.2
dill==0.3.4
diskcache==5.4.0
dlinfo==1.2.1
emoji==1.7.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.3.0/en_core_web_sm-3.3.0-py3-none-any.whl
filelock==3.7.1
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.5.0
ftfy==6.1.1
gensim==4.2.0
huggingface-hub==0.7.0
idna==3.3
isodate==0.6.1
Jinja2==3.1.2
joblib==1.1.0
karateclub==1.2.3
kiwisolver==1.4.2
langcodes==3.3.0
lexicalrichness==0.1.4
MarkupSafe==2.1.1
matplotlib==3.5.2
multidict==6.0.2
multiprocess==0.70.12.2
murmurhash==1.0.7
networkx==2.8.2
nltk==3.7
numpy==1.22.4
packaging==21.3
pandas==1.3.5
pathy==0.6.1
phonemizer==3.2.0
Pillow==9.1.1
preshed==3.0.6
protobuf==4.21.1
pyarrow==8.0.0
pydantic==1.8.2
PyGSP==0.5.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-Levenshtein==0.12.2
python-louvain==0.16
pytz==2022.1
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
responses==0.18.0
rfc3986==1.5.0
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
segments==2.2.0
sentence-transformers==2.2.0
sentencepiece==0.1.96
six==1.16.0
smart-open==5.2.1
spacy==3.3.0
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
stanza==1.4.0
svgling==0.3.1
svgwrite==1.4.2
tabulate==0.8.9
textblob==0.17.1
thinc==8.0.16
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.11.0
tqdm==4.64.0
transformers==4.19.2
typer==0.4.1
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
wasabi==0.9.1
wcwidth==0.2.5
wget==3.2
wincertstore==0.2
xxhash==3.0.0
yarl==1.7.2
zss==1.2.0
